Consolidate exchange testing.
